Modularise conditional show logic
Moved the `meetsCondition` utility function into a `Condition` directory and added in a new utility method for checking that all conditions have been met, which had already been done in two other places (and would have been being done in a third without this refactor).
