+++ src/utils/Condition/index.js
@@ -1,0 +1,9 @@
+import meetsAllConditions from './meetsAllConditions';
+import meetsCondition from './meetsCondition';
+
+const Condition = {
+  meetsAll: meetsAllConditions,
+  met: meetsCondition
+};
+
+export default Condition;
