+++ src/utils/Condition/meetsCondition.js
