+++ src/utils/FormPage/showFormPage.js
@@ -1,24 +1,7 @@
 // Local imports
 import Component from '../Component';
-import Data from '../Data';
-import meetsCondition from '../meetsCondition';
+import Condition from '../Condition';
 import getEditableComponents from './getEditableComponents';
-
-/**
- * Evaluates the show_when condition(s) on a page.
- * @param {object} page The page to consider.
- * @param {object} data The top-level form data.
- * @returns Boolean true if all show_when conditions are met; false otherwise.
- */
-const evaluatePageShowWhen = (page, data) => {
-  const show_when = Array.isArray(page.show_when) ? page.show_when : [page.show_when];
-  let allConditionsMet = true;
-  show_when.forEach(condition => {
-    const sourceDataValue = Data.getSource(data, condition.field);
-    allConditionsMet = allConditionsMet && meetsCondition(condition, sourceDataValue);
-  });
-  return allConditionsMet;
-};
 
 /**
  * Checks whether any of the editable components on a page should be shown.
@@ -45,7 +28,7 @@ const showFormPage = (page, data) => {
 
   // If the page has a show_when condition, we should evaluate that.
   if (page.show_when) {
-    return evaluatePageShowWhen(page, data);
+    return Condition.meetsAll(page.show_when, data);
   }
 
   // If the page itself doesn't have a show_when, we need to make sure that if it
