+++ src/utils/index.js
@@ -4,11 +4,11 @@ import { Utils as HOUtils } from '@ukhomeoffice/cop-react-components';
 // Local imports
 import CheckYourAnswers from './CheckYourAnswers';
 import Component from './Component';
+import Condition from './Condition';
 import Data from './Data';
 import Format from './Format';
 import FormPage from './FormPage';
 import Hub from './Hub';
-import meetsCondition from './meetsCondition';
 import Validate from './Validate';
 
 const Utils = {
@@ -18,7 +18,7 @@ const Utils = {
   Format,
   FormPage,
   Hub,
-  meetsCondition,
+  Condition,
   Validate,
   ...HOUtils
 };
